Converted all floats to double

+++ src/deluge/dsp/delay/delay_buffer.cpp
@@ -134,9 +134,9 @@ void DelayBuffer::setupForRender(int32_t userDelayRate, DelayBufferSetup* setup)
 	if (isResampling) {
 
 		setup->actualSpinRate =
-		    (uint64_t)((float)(userDelayRate << 24) / (float)nativeRate); // 1 is represented as 16777216
+		    (uint64_t)((double)(userDelayRate << 24) / (double)nativeRate); // 1 is represented as 16777216
 		setup->divideByRate =
-		    (uint32_t)((float)0xFFFFFFFF / (float)(setup->actualSpinRate >> 8)); // 1 is represented as 65536
+		    (uint32_t)((double)0xFFFFFFFF / (double)(setup->actualSpinRate >> 8)); // 1 is represented as 65536
 
 		// If buffer spinning slow
 		if (setup->actualSpinRate < 16777216) {
@@ -151,8 +151,8 @@ void DelayBuffer::setupForRender(int32_t userDelayRate, DelayBufferSetup* setup)
 			// because more of that means more "triangle area", or more stuff written each time.
 			//uint32_t delayWriteSizeAdjustment2 = (((uint32_t)delay.speed << 16) / (((uint32_t)(speedMultiple >> 2) * (uint32_t)(speedMultiple >> 2)) >> 11));
 			setup->writeSizeAdjustment =
-			    (uint32_t)((float)0xFFFFFFFF
-			               / (float)(setup->rateMultiple
+			    (uint32_t)((double)0xFFFFFFFF
+			               / (double)(setup->rateMultiple
 			                         * (timesSlowerRead
 			                            + 1))); // Equivalent to one order of magnitude bigger than the above line
 		}
