+++ src/deluge/model/global_effectable/global_effectable_for_clip.cpp
@@ -179,7 +179,7 @@ doNormal:
 			processStutter(globalEffectableBuffer, numSamples, paramManagerForClip);
 
 			int32_t postReverbSendVolumeIncrement =
-			    (int32_t)((float)(postReverbVolume - postReverbVolumeLastTime) / (float)numSamples);
+			    (int32_t)((double)(postReverbVolume - postReverbVolumeLastTime) / (double)numSamples);
 
 			processReverbSendAndVolume(globalEffectableBuffer, numSamples, reverbBuffer, volumePostFX,
 			                           postReverbVolumeLastTime, reverbSendAmount, pan, true,
