+++ src/deluge/model/sample/sample_playback_guide.cpp
@@ -102,10 +102,10 @@ uint64_t SamplePlaybackGuide::getSyncedNumSamplesIn() {
 	}
 	// If following external clock, that could happen
 	// TODO: is that still necessary?
-	return (uint64_t)(((float)(lengthInSamples * currentTickWithinSample)
-	                   + (float)timeSinceLastInternalTick * (float)lengthInSamples / (float)timePerInternalTick
-	                   + (float)(sequenceSyncLengthTicks >> 1)) // Rounding
-	                  / (float)sequenceSyncLengthTicks);
+	return (uint64_t)(((double)(lengthInSamples * currentTickWithinSample)
+	                   + (double)timeSinceLastInternalTick * (double)lengthInSamples / (double)timePerInternalTick
+	                   + (double)(sequenceSyncLengthTicks >> 1)) // Rounding
+	                  / (double)sequenceSyncLengthTicks);
 }
 
 int32_t SamplePlaybackGuide::getNumSamplesLaggingBehindSync(VoiceSample* voiceSample) {
